+++ src/LibraryApplicationSystem.Application/Students/Dto/StudentDto.cs
@@ -11,7 +11,7 @@ namespace LibraryApplicationSystem.Students.Dto
     public class StudentDto : EntityDto<int>
     {
         public string StudentName { get; set; }
-        public int StudentContactNumber { get; set; }
+        public long StudentContactNumber { get; set; }
         public string StudentEmail { get; set; }
         public int DepartmentId { get; set; }
