+++ src/LibraryApplicationSystem.Core/Entities/Student.cs
@@ -6,7 +6,7 @@ namespace LibraryApplicationSystem.Entities
     {
         //Column Name
         public string StudentName { get; set; }
-        public int StudentContactNumber { get; set; }
+        public long StudentContactNumber { get; set; }
         public string StudentEmail { get; set; }
 
         public int DepartmentId { get; set; }
